stopMoving/pyproject.toml
[project]
name = "stopmoving"
version = "0.1.0"
description = ""
authors = [
    {name = "jjjaehong",email = "woghd3014@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "sshtunnel (>=0.4.0,<0.5.0)",
    "requests (>=2.32.4,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "keybert (>=0.9.0,<0.10.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "kiwipiepy (>=0.21.0,<0.22.0)"
]

[tool.poetry]
packages = [{include = "stopmoving", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
